remove the extra code
