+++ package.json
@@ -14,13 +14,6 @@
     "build-dev:scss-to-css": "sass --watch ./webplugin/scss:./webplugin/css/app/style",
     "build:dev": "nodemon ./server/app.js",
     "dev": "npm-run-all -p build:dev build-dev:scss-to-css setup-git-hooks",
-    "test": "npm-run-all build-prod:scss-to-css && NODE_ENV=test node ./server/app.js",
-    "staging": "npm-run-all build-prod:scss-to-css && NODE_ENV=staging node ./server/app.js",
-    "prod": "npm-run-all build-prod:scss-to-css && NODE_ENV=prod node ./server/app.js",
-    "prod-enterprise": "npm-run-all build-prod:scss-to-css && NODE_ENV=prod_enterprise node ./server/app.js",
-    "prod-in": "npm-run-all build-prod:scss-to-css && NODE_ENV=prod_in node ./server/app.js",
-    "prod-eu": "npm-run-all build-prod:scss-to-css && NODE_ENV=prod_eu node ./server/app.js",
-    "prod-cn": "npm-run-all build-prod:scss-to-css && NODE_ENV=prod_cn node ./server/app.js",
     "build-test": "export NODE_ENV=test && npm-run-all build-server gulp-build",
     "build-release": "export NODE_ENV=release && npm-run-all build-server gulp-build",
     "build-prod": "export NODE_ENV=prod && npm-run-all build-server gulp-build",
