+++ server/app.js
@@ -10,7 +10,7 @@ const version = require('../package.json').version;
 const routes = require('./routers/routes.js');
 console.log('Build process started at :', new Date().toString());
 console.log('Version: ', version);
-require('../webplugin/pluginOptimizer.js');
+require('../webplugin/development.js');
 
 console.log('Host url: ', config.urls.hostUrl);
 app.use(cors());
